feat: restructure db configs and schemas
