+++ drizzle.config.ts
@@ -7,7 +7,7 @@ import { defineConfig } from "drizzle-kit";
 config({ path: ".dev.vars" });
 
 export default defineConfig({
-    schema: "./src/lib/db/schema.ts",
+    schema: "./src/db/schema.ts",
     out: "./src/drizzle",
     dialect: "sqlite",
     driver: "d1-http",
