+++ src/db/schema.ts
@@ -1,0 +1,8 @@
+export {
+    account,
+    session,
+    user,
+    verification,
+} from "@/modules/auth/schemas/auth.schema";
+export { categories } from "@/modules/todos/schemas/category.schema";
+export { todos } from "@/modules/todos/schemas/todo.schema";
