+++ src/lib/auth.ts
@@ -3,7 +3,7 @@ import { getCloudflareContext } from "@opennextjs/cloudflare";
 import { betterAuth } from "better-auth";
 import { drizzleAdapter } from "better-auth/adapters/drizzle";
 import { nextCookies } from "better-auth/next-js";
-import { getDb } from "./db";
+import { getDb } from "@/db";
 
 let authInstance: ReturnType<typeof betterAuth> | null = null;
