--- src/lib/db/schema.ts
@@ -1,3 +1,0 @@
-export { account, session, user, verification } from "./schemas/auth.schema";
-export { categories } from "./schemas/category.schema";
-export { todos } from "./schemas/todo.schema";
