+++ src/modules/auth/utils/auth-utils.ts
@@ -4,7 +4,7 @@ import { betterAuth } from "better-auth";
 import { drizzleAdapter } from "better-auth/adapters/drizzle";
 import { nextCookies } from "better-auth/next-js";
 import { headers } from "next/headers";
-import { getDb } from "@/lib/db";
+import { getDb } from "@/db";
 import type { AuthUser } from "@/modules/auth/models/user.model";
 
 /**
