+++ src/modules/todos/actions/update-todo.action.ts
@@ -3,11 +3,14 @@
 import { and, eq } from "drizzle-orm";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
-import { getDb } from "@/lib/db";
-import { todos, updateTodoSchema } from "@/lib/db/schemas/todo.schema";
-import { type TodoPriority, TodoStatus } from "@/lib/enums/todo.enum";
+import { getDb } from "@/db";
 import { type UploadResult, uploadToR2 } from "@/lib/r2";
 import { requireAuth } from "@/modules/auth/utils/auth-utils";
+import {
+    type TodoPriority,
+    TodoStatus,
+} from "@/modules/todos/models/todo.enum";
+import { todos, updateTodoSchema } from "@/modules/todos/schemas/todo.schema";
 import todosRoutes from "../todos.route";
 
 export async function updateTodoAction(todoId: number, formData: FormData) {
