+++ src/modules/todos/schemas/category.schema.ts
@@ -1,8 +1,8 @@
 import { integer, sqliteTable, text } from "drizzle-orm/sqlite-core";
 import { createInsertSchema, createSelectSchema } from "drizzle-zod";
 import { z } from "zod";
-import { VALIDATION_MESSAGES } from "@/lib/constants/validation.constant";
-import { user } from "./auth.schema";
+import { VALIDATION_MESSAGES } from "@/constants/validation.constant";
+import { user } from "@/modules/auth/schemas/auth.schema";
 
 export const categories = sqliteTable("categories", {
     id: integer("id").primaryKey({ autoIncrement: true }),
